19.02.24/b:sprint3/step-1/v:0.1.7.3 react-router lib added; routing started for "constructor" and "404" pages;
