+++ README.md
@@ -57,4 +57,5 @@ a error message common components wrapped by modalOverlay and set it logic;<br>
 v:1.7.0</b>
 
 14.02.24 **b**:sprint3/step-1/**v**: 0.1.7.1
+15.02.24
