+++ package.json
@@ -36,6 +36,7 @@
     "react-dnd": "^16.0.1",
     "react-dnd-html5-backend": "^16.0.1",
     "react-redux": "^9.1.0",
+    "react-router-dom": "^6.22.1",
     "sass": "^1.70.0",
     "typescript": "^5.2.2",
     "uuid": "^9.0.1",
