+++ src/components/AppHeader/AppHeader.tsx
@@ -1,29 +1,32 @@
 import headerStyles from './AppHeaderStyles.module.scss';
 
 import HeaderButton from "./HeaderButton/HeaderButton.tsx";
+import {NavLink} from "react-router-dom";
 
 import {Logo} from "@ya.praktikum/react-developer-burger-ui-components";
 
 export default function AppHeader() {
-    return (
-        <header className={headerStyles.header}>
-
-            <nav className={headerStyles.nav}>
-                <ul>
-                    <li>
-                        <HeaderButton typeFor={"builder"}/>
-                    </li>
-                    <li>
-                        <HeaderButton typeFor={"orders"}/>
-                    </li>
-                </ul>
-                <div className={headerStyles.header_logo}>
-                    <Logo/>
-                </div>
-                <div className={headerStyles.header_profile}>
-                    <HeaderButton typeFor={"profile"}/>
-                </div>
-            </nav>
-        </header>
-    );
+	return (
+		<header className={headerStyles.header}>
+			
+			<nav className={headerStyles.nav}>
+				<ul>
+					<li>
+						<NavLink to="/constructor">
+							<HeaderButton typeFor={"builder"}/>
+						</NavLink>
+					</li>
+					<li>
+						<HeaderButton typeFor={"orders"}/>
+					</li>
+				</ul>
+				<div className={headerStyles.header_logo}>
+					<Logo/>
+				</div>
+				<div className={headerStyles.header_profile}>
+					<HeaderButton typeFor={"profile"}/>
+				</div>
+			</nav>
+		</header>
+	);
 }
