+++ src/pages/BurgerBuilder/BurgerBuilder.module.scss
@@ -1,0 +1,19 @@
+.burger_builder {
+  composes: pl-4 pr-4 from global;
+
+  margin: 0 auto;
+
+  max-width: 1240px;
+  max-height: 80vh;
+
+  display: flex;
+  justify-content: space-between;
+}
+
+.status {
+  position: absolute;
+  left: 50%;
+  top: 50%;
+
+  font-size: 30px;
+}
