+++ src/pages/NotFound404/NotFound404.module.scss
@@ -1,0 +1,3 @@
+.container {
+  background-color: red;
+}
